+++ properties/AddAggregate.php
@@ -15,10 +15,7 @@ use Innmind\BlackBox\{
     Runner\Assert,
 };
 use Innmind\TimeContinuum\Offset;
-use Innmind\Immutable\{
-    Str,
-    Either,
-};
+use Innmind\Immutable\Str;
 use Fixtures\Innmind\TimeContinuum\PointInTime;
 
 /**
@@ -59,11 +56,10 @@ final class AddAggregate implements Property
             ->size();
         $user = User::new($this->createdAt, $this->name);
         $manager->transactional(
-            static fn() => Either::right(
-                $manager
-                    ->repository(User::class)
-                    ->put($user),
-            ),
+            static fn() => $manager
+                ->repository(User::class)
+                ->put($user)
+                ->either(),
         );
         $id = $user->id()->toString();
         unset($user); // to make sure there is no in memory cache somewhere