+++ properties/RemoveSpecification.php
@@ -67,8 +67,8 @@ final class RemoveSpecification implements Property
         $repository = $manager->repository(User::class);
         $manager->transactional(
             static function() use ($repository, $user1, $user2) {
-                $repository->put($user1);
-                $repository->put($user2);
+                $repository->put($user1)->unwrap();
+                $repository->put($user2)->unwrap();
 
                 return Either::right(null);
             },
@@ -83,7 +83,7 @@ final class RemoveSpecification implements Property
                 $repository->remove(Username::of(
                     Sign::equality,
                     Str::of($this->name1),
-                ));
+                ))->unwrap();
 
                 return Either::right(null);
             },