+++ src/Repository.php
@@ -142,8 +142,10 @@ final class Repository
 
     /**
      * @param T $aggregate
+     *
+     * @return Attempt<SideEffect>
      */
-    public function put(object $aggregate): void
+    public function put(object $aggregate): Attempt
     {
         if (!($this->inTransaction)()) {
             throw new \LogicException('Mutation outside of a transaction');
@@ -154,7 +156,7 @@ final class Repository
 
         $this->loaded->add($this, $now);
 
-        $_ = $then->match(
+        return $then->match(
             fn($then) => $this->adapter->update(
                 ($this->diff)($then, $now),
             ),
@@ -198,23 +200,27 @@ final class Repository
 
     /**
      * @param Id<T>|Specification $criteria
+     *
+     * @return Attempt<SideEffect>
      */
-    public function remove(Id|Specification $criteria): void
+    public function remove(Id|Specification $criteria): Attempt
     {
         if (!($this->inTransaction)()) {
             throw new \LogicException('Mutation outside of a transaction');
         }
 
         if (!($criteria instanceof Specification)) {
-            $this->adapter->remove(
-                $this->id->normalize($criteria),
-            );
-            $this->loaded->remove($criteria);
+            return $this
+                ->adapter
+                ->remove($this->id->normalize($criteria))
+                ->map(function($sideEffect) use ($criteria) {
+                    $this->loaded->remove($criteria);
 
-            return;
+                    return $sideEffect;
+                });
         }
 
-        $this->adapter->removeAll(
+        return $this->adapter->removeAll(
             ($this->normalizeSpecification)($criteria),
         );
     }